skeletoned out full board
