+++ src/components/Piece/index.js
@@ -8,10 +8,14 @@ export default function Piece({ name, affiliation }) {
   return (
     <Center
       bg="pastel-tan.0"
-      w="4.5em"
-      h="2.3em"
+      w="4em"
+      h="2em"
       fz="xl"
-      sx={(theme) => ({ borderRadius: '10%', boxShadow: theme.shadows.sm })}>
+      sx={(theme) => ({
+        borderRadius: '10%',
+        boxShadow: theme.shadows.sm,
+        writingMode: 'horizontal-tb'
+      })}>
       {piece.display}
     </Center>
   );
