+++ src/utils.js
@@ -103,6 +103,47 @@ const generateHalfBoardConnections = () => {
 };
 export const halfBoardConnections = generateHalfBoardConnections();
 
+const generateBoardConnections = () => {
+  const connections = [];
+  for (let r = 0; r < 12; r++) {
+    for (let c = 0; c < 4; c++) {
+      connections.push({
+        start: [r, c],
+        end: [r, c + 1]
+      });
+    }
+  }
+  for (let c = 0; c < 5; c++) {
+    for (let r = 0; r < 11; r++) {
+      if (!(r == 5 && (c == 1 || c == 3))) {
+        // no connection through mountains
+        connections.push({
+          start: [r, c],
+          end: [r + 1, c]
+        });
+      }
+    }
+  }
+  const camps = [
+    [2, 1],
+    [2, 3],
+    [4, 1],
+    [4, 3],
+    [7, 1],
+    [7, 3],
+    [9, 1],
+    [9, 3]
+  ];
+  camps.forEach(([r, c]) => {
+    connections.push({ start: [r, c], end: [r - 1, c - 1] });
+    connections.push({ start: [r, c], end: [r - 1, c + 1] });
+    connections.push({ start: [r, c], end: [r + 1, c - 1] });
+    connections.push({ start: [r, c], end: [r + 1, c + 1] });
+  });
+  return connections;
+};
+export const boardConnections = generateBoardConnections();
+
 export const isHalfBoardHQ = (r, c) => r === 5 && (c === 1 || c === 3);
 
 export const isHalfBoardRailroad = (r, c) => {
@@ -110,6 +151,13 @@ export const isHalfBoardRailroad = (r, c) => {
     return false;
   }
   return r === 0 || r === 4 || c === 0 || c === 4;
+};
+
+export const isRailroad = (r, c) => {
+  if (c === 0 || c === 4) {
+    return r >= 1 && r <= 10;
+  }
+  return r === 1 || r === 5 || r === 6 || r === 10;
 };
 
 export const isValidHalfBoardPlacement = (piece, row, col) => {
@@ -135,3 +183,13 @@ export const isValidHalfBoardPlacement = (piece, row, col) => {
 
   return true;
 };
+
+export const emptyBoard = () => {
+  const board = [];
+  for (let i = 0; i < 12; i++) {
+    board.push(Array(5).fill(null));
+  }
+  return board;
+};
+
+export const isHQ = (r, c) => (c == 1 || c == 3) && (r == 0 || r == 11);
